Hold transport protocol number in PacketBuffer
There are several places where we need to find out what kind of
transport protocol header is in the header storage and we don't want to
have to parse the IPV6 extension headers again just to find out. Generally
speaking we are not yet using it on output but make sure it is valid as
we will probably use it in the filtering. This is a preparatory CL. It
gets used in a following change. In a couple of
the tests here we are just symbolically setting the value to 0. (no
specific transport header) just to be consistent in setting it whenever
we set a transport header into the packet Buffer.

Updates #2211.

PiperOrigin-RevId: 328635575